requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
tqdm
matplotlib
pillow
opencv-python
pyyaml